podcast_intel_agent/env_bootstrap.py
"""Load project-root `.env` before other modules read `os.environ`."""

from pathlib import Path

from dotenv import load_dotenv

_PROJECT_ROOT = Path(__file__).resolve().parent.parent
load_dotenv(_PROJECT_ROOT / ".env")
