requirements.txt
google-adk>=1.0.0
feedparser>=6.0.0
openai-whisper>=20231117
torch>=2.0.0
python-dotenv>=1.0.0
openai>=1.40.0
